code/requirements.txt
sentence-transformers
torch
transformers
llama-cpp-python
grpcio
grpcio-tools
fastapi
uvicorn
pydantic
